checkpoint: initial schema for tracking resolver records
